# authors_app/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Authors API",
    "description": "API for managing authors, books, and companies"
  },
  "basePath": "/api/v1",
  "paths": {
    "/auth/register": {
      "post": {
        "summary": "Register a new user",
        "description": "Register a new user with the provided information",
        "parameters": [
          {
            "name": "User",
            "in": "body",
            "description": "User details",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "first_name": { "type": "string" },
                "last_name": { "type": "string" },
                "email": { "type": "string" },
                "image": { "type": "string" },
                "biography": { "type": "string" },
                "user_type": { "type": "string" },
                "password": { "type": "string" },
                "contact": { "type": "string" }
              }
            }
          }
        ],
        "responses": {
          "201": { "description": "User created successfully" },
          "400": { "description": "Bad request" },
          "409": {
            "description": "Conflict - User with this email already exists"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Login",
        "description": "Login with email and password",
        "parameters": [
          {
            "name": "Credentials",
            "in": "body",
            "description": "User email and password",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": { "type": "string" },
                "password": { "type": "string" }
              }
            }
          }
        ],
        "responses": {
          "200": { "description": "Login successful" },
          "401": { "description": "Unauthorized - Invalid email or password" }
        }
      }
    },
    "/auth/edit/{id}": {
      "put": {
        "summary": "Edit user details",
        "description": "Edit user details by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User ID",
            "required": true,
            "type": "integer"
          },
          {
            "name": "User",
            "in": "body",
            "description": "Updated user details",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "first_name": { "type": "string" },
                "last_name": { "type": "string" },
                "email": { "type": "string" },
                "image": { "type": "string" },
                "biography": { "type": "string" },
                "user_type": { "type": "string" },
                "password": { "type": "string" },
                "contact": { "type": "string" }
              }
            }
          }
        ],
        "responses": {
          "201": { "description": "User updated successfully" },
          "404": { "description": "Not found - User not found" },
          "500": { "description": "Internal Server Error" }
        }
      }
    },
    "/auth/delete/{id}": {
      "delete": {
        "summary": "Delete user",
        "description": "Delete user by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User ID",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": { "description": "User deleted successfully" },
          "404": { "description": "Not found - User not found" },
          "500": { "description": "Internal Server Error" }
        }
      }
    },
    "/auth/user/{id}": {
      "get": {
        "summary": "Get specific user",
        "description": "Get user details by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User ID",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": { "description": "Success" },
          "404": { "description": "Not found - User not found" },
          "500": { "description": "Internal Server Error" }
        }
      }
    },
    "/auth/users": {
      "get": {
        "summary": "Get all users",
        "description": "Get details of all users",
        "responses": {
          "200": { "description": "Success" },
          "500": { "description": "Internal Server Error" }
        }
      }
    },
    "/books/create": {
      "post": {
        "summary": "Create a new book",
        "description": "Create a new book with the provided information",
        "parameters": [
          {
            "name": "Book",
            "in": "body",
            "description": "Book details",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "title": { "type": "string" },
                "description": { "type": "string" },
                "price": { "type": "number" },
                "image": { "type": "string" },
                "pages": { "type": "integer" },
                "company_id": { "type": "integer" },
                "isbn": { "type": "string" },
                "genre": { "type": "string" }
              }
            }
          }
        ],
        "responses": {
          "201": { "description": "Book created successfully" },
          "400": { "description": "Bad request" }
        }
      }
    },
    "/companies/create": {
      "post": {
        "summary": "Create a new company",
        "description": "Create a new company with the provided information",
        "parameters": [
          {
            "name": "Company",
            "in": "body",
            "description": "Company details",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": { "type": "string" },
                "description": { "type": "string" },
                "address": { "type": "string" },
                "website": { "type": "string" },
                "logo": { "type": "string" }
              }
            }
          }
        ],
        "responses": {
          "201": { "description": "Company created successfully" },
          "400": { "description": "Bad request" }
        }
      }
    }
  }
}